backend/requirements.txt

langchain>=0.2
langchain-openai>=0.1
langgraph>=0.1.6
fastapi>=0.111
uvicorn>=0.30
sse-starlette>=2.1.0
pydantic>=2.7
python-dotenv>=1.0
